requirements.txt
azure-cognitiveservices-knowledge-qnamaker==0.2.0
azure-common==1.1.25
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.9
idna==2.10
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
msrest==0.6.19
oauthlib==3.1.0
python-dotenv==0.14.0
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
urllib3==1.25.10
Werkzeug==1.0.1
